Change to vite and update js -> jsx
