+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/hona_256.ico" />
+    <link rel="icon" href="/hona_256.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta name="description" content="Hemsida för Hensmåla Triathlon" />
@@ -10,11 +10,12 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,400;0,700;0,900;1,300&display=swap" rel="stylesheet">
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <link rel="manifest" href="/manifest.json" />
     <script
       src="https://kit.fontawesome.com/1f5f856382.js"
       crossorigin="anonymous"
     ></script>
+    <script type="module" src="/src/index.jsx"></script>
     <script src="https://www.google.com/recaptcha/api.js?render=6LcKIqQZAAAAAK88TdJkAsZAOZ4YLSf7VFqtXMNz"></script>
