+++ netlify.toml
@@ -1,0 +1,2 @@
+[dev]
+  framework = "vite"
