+++ package.json
@@ -4,6 +4,8 @@
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.11.4",
+    "@netlify/functions": "^2.8.2",
+    "@vitejs/plugin-react": "^4.3.3",
     "axios": "^1.6.8",
     "bootstrap": "^4.6.0",
     "dotenv": "^8.2.0",
@@ -22,15 +24,14 @@
     "react-spinners": "^0.13.8",
     "react-youtube": "^10.1.0",
     "reactstrap": "8.9.0",
-    "slick-carousel": "^1.8.1"  },
-  "devDependencies": {
-    "react-scripts": "^5.0.1"
+    "slick-carousel": "^1.8.1",
+    "vite": "^5.4.10",
+    "vite-tsconfig-paths": "^5.0.1"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "vite",
+    "build": "tsc && vite build",
+    "serve": "vite preview"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -46,5 +47,10 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "typescript": "^5.6.3"
   }
 }
