+++ src/components/about/GeneralInfo.jsx
