+++ src/components/about/Rules.jsx
