+++ src/components/challenge/EntryCard.jsx
