+++ src/components/challenge/EntryModal.jsx
