+++ src/components/news/NewsDetail.jsx
