+++ src/components/register/RegisterFormSolo.jsx
@@ -6,44 +6,30 @@ import {
   Input,
   Row,
   Col,
-  Card,
-  CardBody,
   FormText
 } from "reactstrap";
 import { NavLink as RRNavLink } from "react-router-dom";
 import Consent from "../Consent";
-import ShirtSelect from "./ShirtSelect";
-import CapSelect from "./CapSelect";
 import ExtraDonation from "./ExtraDonation";
 import { DayPicker, MonthPicker, YearPicker } from "../TimeUtils";
+import ShirtSelect from "./ShirtSelect";
+import CapSelect from "./CapSelect";
 import RegisterButton from "./RegisterButton";
-import { scrollToInfo, calcShirtPrice, SHIRT_PRICE_COTTON, SHIRT_PRICE_FUNCTIONAL, CAP_PRICE} from './Utils';
+import { scrollToInfo, calcShirtPrice, SHIRT_PRICE_COTTON, SHIRT_PRICE_FUNCTIONAL, CAP_PRICE } from './Utils';
 
-const LATE_REGISTER_FEE = 700;
+const LATE_REGISTER_FEE = 400;
 const REGISTER_FEE = LATE_REGISTER_FEE;
 
-class RegisterFormTeam extends Component {
+class RegisterFormSolo extends Component {
   state = {
-    teamName: "",
-    name1: "",
-    email1: "",
-    year1: "",
-    month1: "",
-    day1: "",
-    city1: "",
-    name2: "",
-    email2: "",
-    year2: "",
-    month2: "",
-    day2: "",
-    city2: "",
-    name3: "",
-    email3: "",
-    year3: "",
-    month3: "",
-    day3: "",
-    city3: "",
+    name: "",
+    email: "",
+    year: "",
+    month: "",
+    day: "",
     info: "",
+    gender: "",
+    city: "",
     isCheckboxOneTicked: false,
     isCheckboxTwoTicked: false,
     isCheckboxThreeTicked: false,
@@ -74,10 +60,7 @@ class RegisterFormTeam extends Component {
   };
 
   isAllowedCompanyEntered = () => {
-    return process.env.REACT_APP_ALLOWED_COMPANY && 
-      (this.state.city1.toLowerCase().includes(process.env.REACT_APP_ALLOWED_COMPANY.toLowerCase()) ||
-      this.state.city2.toLowerCase().includes(process.env.REACT_APP_ALLOWED_COMPANY.toLowerCase()) ||
-      this.state.city3.toLowerCase().includes(process.env.REACT_APP_ALLOWED_COMPANY.toLowerCase()));
+    return import.meta.env.REACT_APP_ALLOWED_COMPANY && this.state.city.toLowerCase().includes(import.meta.env.REACT_APP_ALLOWED_COMPANY.toLowerCase());
   }
 
   calcTotalCost = () => {
@@ -87,89 +70,18 @@ class RegisterFormTeam extends Component {
     return REGISTER_FEE + this.state.extraDonation + calcShirtPrice(this.state.shirts) + this.state.numCaps * CAP_PRICE;
   }
 
-  renderMemberFields = () => {
-    return [1, 2, 3].map((num) => {
-      return(
-      <div key={num} style={{marginBottom: "20px"}}>
-        <Label for={`teamMember${num}`}>Lagmedlem {num}</Label>
-        {num === 3 ? <i> (Ej för lag med endast två deltagare)</i> : null} 
-        <i></i>
-          <Card id={`teamMember${num}`} style={{ backgroundColor: `#dfeff${num*3}` }}>
-            <CardBody>
-              <FormGroup>
-                <Label for={`name${num}`}>Namn*</Label>
-                <Input
-                  required={num !== 3}
-                  type="text"
-                  name={`name${num}`}
-                  id={`name${num}`}
-                  placeholder="Förnamn Efternamn"
-                  value={this.state[`name${num}`]}
-                  onChange={this.handleChange}
-                />
-              </FormGroup>
-              <FormGroup>
-                <Label for={`email${num}`}>Epost*</Label>
-                <Input
-                  required={num !== 3}
-                  type="email"
-                  name={`email${num}`}
-                  id={`email${num}`}
-                  placeholder="din.email@gmail.com"
-                  value={this.state[`email${num}`]}
-                  onChange={this.handleChange}
-                />
-              </FormGroup>
-              <FormGroup>
-                <Label for={`birthdayID${num}`}>Födelsedatum*</Label>
-                <div style={{ display: "flex" }}>
-                  <YearPicker
-                    required={num !== 3}
-                    handleChange={this.handleChange}
-                    elemName={`year${num}`}
-                  />
-                  <MonthPicker
-                    required={num !== 3}
-                    handleChange={this.handleChange}
-                    elemName={`month${num}`}
-                  />
-                  <DayPicker
-                    required={num !== 3}
-                    handleChange={this.handleChange}
-                    elemName={`day${num}`}
-                  />
-                </div>
-              </FormGroup>
-              <FormGroup>
-                <Label for={`city${num}`}> Ort (klubb)</Label>
-                <Input
-                  type="text"
-                  name={`city${num}`}
-                  id={`city${num}`}
-                  placeholder="Hensmåla löparförening"
-                  value={this.state[`city${num}`]}
-                  onChange={this.handleChange}
-                />
-              </FormGroup>
-            </CardBody>
-          </Card>
-        </div>
-        )
-    })
-  }
-            
-
   render() {
     return (
       <Row>
         <Col style={{ marginTop: "2vh" }} md={6}>
           <Form
             onSubmit={(e) =>
-              this.props.handleSubmit(e, "team", this.state, this.calcTotalCost())
+              this.props.handleSubmit(e, "solo", this.state, this.calcTotalCost())
             }
           >
             <div style={{ display: "flex", justifyContent: "space-between" }}>
-              <h3>Anmälan 2024 Lag</h3>
+              <h3>Anmälan 2024 Individuell</h3>
+              
               <div onClick={() => scrollToInfo("info-text")} className="scroll-to-info-btn">
                 Visa info<i className="fas fa-angle-down angle-down"></i>
               </div>
@@ -180,18 +92,83 @@ class RegisterFormTeam extends Component {
               </b>
             </p>
             <FormGroup>
-              <Label for="teamName">Lagnamn*</Label>
+              <Label for="name">Namn*</Label>
               <Input
                 required={true}
                 type="text"
-                name="teamName"
-                id="teamName"
-                placeholder="Gubbaflås"
-                value={this.state.teamName}
+                name="name"
+                id="name"
+                placeholder="Förnamn Efternamn"
+                value={this.state.name}
                 onChange={this.handleChange}
               />
             </FormGroup>
-            {this.renderMemberFields()}
+
+            <FormGroup>
+              <Label for="email">Epost*</Label>
+              <Input
+                required={true}
+                type="email"
+                name="email"
+                id="email"
+                placeholder="din.email@gmail.com"
+                value={this.state.email}
+                onChange={this.handleChange}
+              />
+            </FormGroup>
+            <FormGroup>
+              <Label for="birthdayID">Födelsedatum*</Label>
+              <div style={{ display: "flex" }}>
+                <YearPicker 
+                  reqired={true} 
+                  handleChange={this.handleChange} 
+                  elemName="year" />
+                <MonthPicker
+                  reqired={true}
+                  handleChange={this.handleChange}
+                  elemName="month"
+                />
+                <DayPicker 
+                reqired={true} 
+                handleChange={this.handleChange} 
+                elemName="day" />
+              </div>
+            </FormGroup>
+            <FormGroup>
+              <Label for="genderSelection">Kön*</Label>
+              <Input
+                required={true}
+                type="select"
+                name="gender"
+                id="genderSelection"
+                onChange={this.handleChange}
+              >
+                <option disabled selected value>
+                  Välj kön
+                </option>
+                <option value="Man">Man</option>
+                <option value="Kvinna">Kvinna</option>
+              </Input>
+            </FormGroup>
+
+            <FormGroup>
+              <Label for="city"> Ort (klubb)</Label>
+              <Input
+                type="text"
+                name="city"
+                id="city"
+                placeholder="Hensmåla löparförening"
+                value={this.state.city}
+                onChange={this.handleChange}
+              />
+            </FormGroup>
+            {this.isAllowedCompanyEntered() ? 
+            <div className="allowed-company-text-bg">
+              <small>
+              Du har anget <b style={{color: "#007fa8"}}>{import.meta.env.REACT_APP_ALLOWED_COMPANY}</b> som klubb och får därför anmälningsavgiften betald.
+              </small>
+            </div>
+            : null}
             <FormGroup>
               <Label for="info">Information</Label>
               <Input
@@ -204,14 +181,14 @@ class RegisterFormTeam extends Component {
               />
             </FormGroup>
             <FormGroup>
-                <Label for="clothes-select">Lägg till t-shirt (Bomull {SHIRT_PRICE_COTTON}kr, Funktion {SHIRT_PRICE_FUNCTIONAL}kr)</Label>
-                <div className="clothes-select">
-                  <ShirtSelect updateShirtSelection={(newShirts) => this.setState({shirts: newShirts})}/>
-                </div>
-                <Label className="mt-2">Lägg till keps ({CAP_PRICE}kr)</Label>
-                <div className="clothes-select">
-                  <CapSelect updateCapSelection={(numCaps) => this.setState({numCaps: numCaps})}/>
-                </div>
+                <Label>Lägg till t-shirt (Bomull {SHIRT_PRICE_COTTON}kr, Funktion {SHIRT_PRICE_FUNCTIONAL}kr)</Label>
+              <div className="clothes-select">
+                <ShirtSelect updateShirtSelection={(newShirts) => this.setState({shirts: newShirts})}/>
+              </div>
+              <Label className="mt-2">Lägg till keps ({CAP_PRICE}kr)</Label>
+              <div className="clothes-select">
+                <CapSelect updateCapSelection={(numCaps) => this.setState({numCaps: numCaps})}/>
+              </div>
             </FormGroup>
             <FormGroup>
               <Label for="extra-donation">Extra donation till ALS-forskningen</Label>
@@ -221,9 +198,9 @@ class RegisterFormTeam extends Component {
               <FormText color="bold">* obligatoriska fält.</FormText>
             </FormGroup>
             <FormGroup check>
-              <Label for="checkbox1Team" className="consent-checkbox">
+              <Label for="checkbox1" className="consent-checkbox">
                 <Input
-                  id="checkbox1Team"
+                  id="checkbox1"
                   type="checkbox"
                   onClick={() => this.toggleConsent(1)}
                 />{" "}
@@ -235,9 +212,9 @@ class RegisterFormTeam extends Component {
               />
             </FormGroup>
             <FormGroup check>
-              <Label for="checkbox2Team">
+              <Label for="checkbox2">
                 <Input
-                  id="checkbox2Team"
+                  id="checkbox2"
                   type="checkbox"
                   onClick={() => this.toggleConsent(2)}
                 />{" "}
@@ -253,9 +230,9 @@ class RegisterFormTeam extends Component {
               </Label>
             </FormGroup>
             <FormGroup check>
-              <Label for="checkbox3Team">
+              <Label for="checkbox3">
                 <Input
-                  id="checkbox3Team"
+                  id="checkbox3"
                   type="checkbox"
                   onClick={() => this.toggleConsent(3)}
                 />{" "}
@@ -263,29 +240,21 @@ class RegisterFormTeam extends Component {
                 internet.
               </Label>
             </FormGroup>
-            {this.isAllowedCompanyEntered() ? 
-            <div className="allowed-company-text-bg">
-              <small>
-              Du har anget <b style={{color: "#007fa8"}}>{process.env.REACT_APP_ALLOWED_COMPANY}</b> som klubb och får därför anmälningsavgiften betald.
-              </small>
-            </div>
-            : null}
             <FormGroup>
               <Label for="totalAmountToPay">Totalt att betala:</Label>
               <h5>{this.calcTotalCost()}kr</h5>
             </FormGroup>
-            <RegisterButton 
-              id="submitButton"
-              text="Anmäl oss!" 
-              disabled={
-                !(
-                  this.state.isCheckboxOneTicked &&
-                  this.state.isCheckboxTwoTicked &&
-                  this.state.isCheckboxThreeTicked
-                ) || this.props.loading
-              }
-              loading={this.props.loading}
-            />
+          <RegisterButton
+            text="Anmäl mig!" 
+            disabled={
+              !(
+                this.state.isCheckboxOneTicked &&
+                this.state.isCheckboxTwoTicked &&
+                this.state.isCheckboxThreeTicked
+              ) || this.props.loading
+            }
+            loading={this.props.loading}
+          />
           </Form>
           <small>
             This site is protected by reCAPTCHA and the Google{" "}
@@ -295,13 +264,16 @@ class RegisterFormTeam extends Component {
           </small>
         </Col>
         <Col id="info-text" style={{ marginTop: "2vh" }}>
-          <hr className="register-divider"></hr>          
-          <h3>Anmäl er som Lag (2-3 pers.)</h3>
+          <hr className="register-divider"></hr>
+          <h3>Anmäl dig till Hensmåla Triathlon 2024</h3>
           <b>Datum: 20 juli</b>
           <p>
-            När ni anmäler er som lag får sträckorna delas upp hur ni vill inom laget. Detta
-            kan innebära att ni är tre personer som deltar där alla kör en gren var, eller ett lag med 2 personer där en av er kör 2 grenar. För mer information om sträckorna och tävlingsregler kan
-            du gå in{" "}
+            Ett motionslopp för alla, motionär som elit. Få en härlig dag i Hensmålas vackra natur med simning (340m), cykling (9.2km), löpning (6.5km) och samtidigt bidra till ALS-forskningen.
+          </p>
+          <p>
+            När du anmälder sig som individuell deltagare utför du alla tre
+            grenar på egen hand. För mer information om sträckorna och
+            tävlingsregler kan du gå in{" "}
             <RRNavLink 
               target="_blank"
               rel="noopener noreferrer" 
@@ -322,7 +294,7 @@ class RegisterFormTeam extends Component {
             vara med.
           </b>
           <br></br>
-          <br></br>   
+          <br></br>
           <p>Vid frågor kontakta hensmala.triathlon@gmail.com</p>
           <b style={{ fontSize: 20 }}>Startavgift: {REGISTER_FEE}kr</b>
         </Col>
@@ -331,4 +303,4 @@ class RegisterFormTeam extends Component {
   }
 }
 
-export default RegisterFormTeam;
+export default RegisterFormSolo;
